Data.Factual.Credentials

Description

This module contains a simple definition of the credentials type which is used to perform submit queries to the Factual API.

Credentials type

data Credentials

A Credentials value is used to generate a Token.

Constructors

Credentials Key Secret

type Key = String

The Key is the oauth key as a String.

type Secret = String

The Secret is the oauth secret as a String.