Data.Factual.Query

Description

This module exports the definition of the Query typeclass.

Query typeclass

class Query q where

A member of the Query typeclass must define a toPath method which converts the Query into a path String.

Methods

toPath ∷ q → String