Data.Factual.Query.CrosswalkQuery

Description

This module exports the type used to create crosswalk queries.

CrosswalkQuery type

data CrosswalkQuery

A crosswalk query can be formed by specifying a factual id and (optionally) a list of namespaces to only include when attempting to find the ids for various namespaces, or by specifying a namespace and namespace id in order to find the factual id. An optional limit can be set as well.

Constructors

CrosswalkQuery

Fields

factualId :: Maybe String

limit :: Maybe Int

namespace :: Maybe String

namespaceId :: Maybe String

only :: [String]