Data.Factual.Query.DiffsQuery

Description

This module exports the types used to create diffs queries.

DiffsQuery type

data DiffsQuery

The DiffsQuery type is used to construct diffs queries. A table, start timestamp and end timestamp should be specified.

Constructors

DiffsQuery

Fields

table ∷ Table

start ∷ Timestamp

end ∷ Timestamp

Required modules

module Data.Factual.Shared.Table