Data.Factual.Query.FacetsQuery

Description

This module exports the types used to create facets queries.

FacetsQuery type

data FacetsQuery

The FacetsQuery type is used to construct facets queries. A table and search should be specified, but the rest of the query options are essentially optional.

Constructors

FacetsQuery

Fields

table :: Table

search :: Search

select :: [String]

filters :: [Filter]

geo :: Maybe Geo

limit :: Maybe Int

minCount :: Maybe Int

includeCount :: Bool

module Data.Factual.Shared.Table

module Data.Factual.Shared.Search

module Data.Factual.Shared.Filter

module Data.Factual.Shared.Geo