Data.Factual.Query.GeopulseQuery

Description

This module exports the type used to create geopulse queries.

ResolveQuery type

data GeopulseQuery

The GeopulseQuery type is used to construct geopulse queries. A geo point is required but select values are optional (just use an empty list to denote selecting all pulses).

Constructors

GeopulseQuery

Fields

geo ∷ Geo

select ∷ [String]

Required modules

module Data.Factual.Shared.Geo