Data.Factual.Query.MatchQuery

Description

This module exports the type used to create match queries.

MatchQuery type

data MatchQuery

A match query is formed as an array of match values. These values will be compared with Factual records to return a cleaner, more canonical row of data.

Constructors

MatchQuery [MatchValue]

MatchValue type

data MatchValue

A match value can either be a String or a Number (Double). The first argument is the name of the field and the second argument is the input value.

Constructors

MatchStr String String

MatchNum String Double