Data.Factual.Query.ResolveQuery

Description

This module exports the type used to create resolve queries.

ResolveQuery type

data ResolveQuery

A resolve query is formed as an array of resolve values. These values will be compared with Factual records to return a cleaner, more canonical row of data.

Constructors

ResolveQuery

Fields

values ∷ [ResolveValue]

debug ∷ Bool

ResolveValue type

data ResolveValue

A resolve value can either be a String or a Number (Double). The first argument is the name of the field and the second argument is the input value.

Constructors

ResolveStr String String

ResolveNum String Double