Data.Factual.Query.SchemaQuery

Description

This module exports the type used to create schema queries.

SchemaQuery type

data SchemaQuery

A schema query is formed by simply supplying a Table to the value constructor.

Constructors

SchemaQuery Table

Required modules

module Data.Factual.Shared.Table