Data.Factual.Shared.Geo

Description

This module exports the Geo type used to create read and facet queries.

Geo type

type Lat = Double

A Lat is the latitude represented as a Double.

type Long = Double

A Long is the longitude represented as a Double.

type Radius = Double

A Radius is the radius of the circle as a Double in meters.

data Geo

The Geo type is used to limit the search to specific geograph location. Currently, only circles are supported. Supply a latitude, longitude and radius in meters for the circle.

Constructors

Circle Lat Long Radius

Point Lat Long

Helper functions

geoString :: Maybe Geo -> String