Data.Factual.Shared.Table

Description

This module exports the type used to represent a table for the read or schema query types.

Table type

data Table

This type defines the available tables. Use the Custom table option for tables that are not listed you.

Constructors

Places

RestaurantsUS

HotelsUS

Global

Crosswalk

HealthCareProviders

WorldGeographies

ProductsCPG

ProductsCrosswalk

Monetize

Custom String