Data.Factual.Utils

Description

This module exports shared utility methods used by the Factual datatypes.

Utility methods

join ∷ [a] → [[a]] → [a]

The join function joins a list of lists into a list using a separator list. The most common use case is for joining Strings with a common separator String.

joinAndFilter ∷ [String] → String

This function filters out empty Strings from a list before joining the Strings with an & character. The use case is forming query path Strings.

selectString ∷ [String] → String

limitString ∷ Maybe Int → String

includeCountString ∷ Bool → String