Data.Factual.Write

Description

This module exports the definition of the Write typeclass.

Query typeclass

class Write w where

A member of the Write typeclass must define a url method which returns the write url as a string and a body method which returns of the body of the write.

Methods

path :: w -> String

body :: w -> String