Data.Factual.Write.Flag

Description

This module exports the types used to create flag writes.

Flag type

data Flag

The Flag type represents a Write to be made to the API which flags a row as having some kind of problem. The table and factualId identify the problematic row, while the problem indicates the type of issue the row has. The user is specified as a string. Other fields such as comment and reference are optional. The debug flag is used to write in debug mode.

Constructors

Flag

Fields

table :: Table

factualId :: String

problem :: Problem

user :: String

comment :: Maybe String

debug :: Bool

reference :: Maybe String

Problem type

data Problem

A Problem represents what is wrong with the row being flagged

Constructors

Duplicate

Nonexistent

Inaccurate

Inappropriate

Spam

Other

Required modules

module Data.Factual.Shared.Table