Data.Factual.Write.Submit

Description

This module exports the types used to create contributions.

Submit type

data Submit

The Submit type represents a Write to the API which performs an upsert (a row can be updated or a new row can be written). The table and user must be specified, while the factual ID is optional (omitted for new rows). Finally the values are specified in a String to String Map.

Constructors

Submit

Fields

table :: Table

user :: String

factualId :: Maybe String

values :: Map String String

Required modules

module Data.Factual.Shared.Table