Network.Factual.API

Description

This module exports functions which are used to execute requests and handle the OAuth authentication process.

Authentication

generateToken ∷ Key → Secret → Token

This function takes a Key and Secret, and generates a Token that is passed to the various methods used to make requests.

Read functions

executeQuery ∷ Query query ⇒ Token → query → IO Response

This function takes a Token and Query value and sends the query to the Factual API. The resultant IO action contains a Response value which wraps the resultant data.

executeMultiQuery ∷ Query query ⇒ Token → Map String query → IO (Map String Response)

This function can be used to make a Multi Query (multiple queries in a single request. It takes a Token, a Map of key Strings to Queries and returns a Map from the same keys to Response values.

get ∷ Token → Path → Params → IO Response

This function can be used to perform raw read queries to any API endpoint. It takes a Token, a Path string and a Map of params (both keys and values are strings). The function returns a standard Response value.

Write functions

executeWrite ∷ Write write ⇒ Token → write → IO Response

This function is used to execute Writes. The function takes a Token and a Write value, and returns a Response value.

post ∷ Token → Path → Params → Body → IO Response

This function can be used to perform raw post queries to any API endpoint. It takes a Token, a Path string, a Map of params and a body Map. Both Maps have String keys and values. The function returns a standard Response value.

Debug functions

debugQuery ∷ Query query ⇒ query → IO ()

This function can be used to debug Queries. It takes a Query value and prints out the URL path generated by that query.

debugWrite ∷ Write write ⇒ write → IO ()

This function can be used to debug Writes. It takes a Write value and prints out the URL path, and post body generated by that write.

The hoauth Token type

data Token

Constructors

TwoLegg

Fields

application ∷ Application

oauthParams ∷ FieldList

ReqToken

Fields

application ∷ Application

oauthParams ∷ FieldList

AccessToken

Fields

application ∷ Application

oauthParams ∷ FieldList