Network.Factual.API

Description

This module exports functions which are used to execute queries and handle the OAuth authentication process.

Authentication

generateToken :: Key -> Secret -> Token

This function takes a set of credentials and returns an OAuth token that can be used to make requests.

Read functions

makeRequest :: Query query => Token -> query -> IO Response

This function takes an OAuth token and a query (which is member of the Query typeclass) and returns an IO action which will fetch a response from the Factual API.

makeRawRequest :: Token -> String -> IO Response

This function can be used to make raw read requests for any path. You pass in your Token and the path of your request (e.g. "/t/places?q=starbucks")

Write functions

sendWrite :: Write write => Token -> write -> IO Response

This function takes an OAuth token and a Write and retunrs and IO action which sends the Write to API and returns a Response.

The hoauth Token type

data Token

Constructors

TwoLegg

Fields

application :: Application

oauthParams :: FieldList

ReqToken

Fields

application :: Application

oauthParams :: FieldList

AccessToken

Fields

application :: Application

oauthParams :: FieldList